requirements.txt

django==4.0.0

# for admin

django-modeladmin-reorder
django-tinymce
django-import-export
django-taggit
django-nested-admin

# for apis
djangorestframework
django-cors-headers
django-filter

